@@ -1,4 +1,4 @@
-// swift-tools-version:5.5
+// swift-tools-version:5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,14 +6,16 @@ import PackageDescription
 let package = Package(
     name: "RSWTools",
     platforms: [
-        .iOS(.v13),
+        .iOS(.v15),
+        .visionOS(.v1)
     ], products: [
         .library(
             name: "RSWTools",
-            targets: ["RSWTools"]),
+            targets: ["RSWTools"]
+        ),
     ],
     dependencies: [
-        .package(name: "SwiftTools", url: "https://github.com/Rspoon3/SwiftTools.git", from: "1.0.0")
+        .package(url: "https://github.com/Rspoon3/SwiftTools.git", exact: "1.1.0")
     ],
     targets: [
         .target(
@@ -24,6 +26,7 @@ let package = Package(
         ),
         .testTarget(
             name: "RSWToolsTests",
-            dependencies: ["RSWTools"]),
+            dependencies: ["RSWTools"]
+        ),
     ]
 )
