@@ -9,13 +9,11 @@ import SwiftUI
 import SFSymbols
 import SwiftTools
 
-
-@available(iOS 14, *)
 public struct SFSymbolPicker: View {
-    @Environment(\.presentationMode) var presentationMode
-    @StateObject var searchBar = SearchBar()
+    @Environment(\.dismiss) var dismiss
     @State private var currentSymbols = SFSymbol.allSymbols
     @State private var sort: SFCategory = .all
+    @State private var searchText = ""
     
     public var symbolTitle: String
     public var result: (SFSymbol)->Void
@@ -23,17 +21,22 @@ public struct SFSymbolPicker: View {
     
     let columns = [GridItem(.adaptive(minimum: 65))]
     
-    var searchResults: [SFSymbol]{
-        if searchBar.text.isEmpty{
+    var searchResults: [SFSymbol] {
+        guard !searchText.isEmpty else {
             return currentSymbols
         }
         
-        return currentSymbols.filter{$0.title.localizedCaseInsensitiveContains(searchBar.text)}
+        return currentSymbols.filter{
+            $0.title.localizedCaseInsensitiveContains(searchText)
+        }
     }
     
-    
     //MARK: Initializer
-    public init(symbolTitle: String, color: Color = .accentColor, result: @escaping (SFSymbol) -> Void) {
+    public init(
+        symbolTitle: String,
+        color: Color = .accentColor,
+        result: @escaping (SFSymbol) -> Void
+    ) {
         self.symbolTitle = symbolTitle
         self.result = result
         self.color = color
@@ -45,8 +48,9 @@ public struct SFSymbolPicker: View {
         ZStack{
             Color(.systemGroupedBackground)
                 .edgesIgnoringSafeArea(.all)
-            ScrollView{
-                LazyVGrid(columns: columns, spacing: 20){
+            
+            ScrollView {
+                LazyVGrid(columns: columns, spacing: 20) {
                     ForEach(searchResults) { symbol in
                         ZStack(alignment: .topTrailing){
                             Image(symbol: symbol)
@@ -57,13 +61,13 @@ public struct SFSymbolPicker: View {
                                 .foregroundColor(color)
                                 .onTapGesture{
                                     result(symbol)
-                                    presentationMode.wrappedValue.dismiss()
+                                    dismiss()
                                 }
                                 .contextMenu{
                                     Text(symbol.title)
                                 }
                             
-                            if symbolTitle == symbol.title{
+                            if symbolTitle == symbol.title {
                                 Image(symbol: .checkmark)
                                     .resizable()
                                     .aspectRatio(contentMode: .fit)
@@ -79,10 +83,16 @@ public struct SFSymbolPicker: View {
             .navigationTitle("Symbol Picker")
             .navigationBarTitleDisplayMode(.inline)
             .animation(.default, value: searchResults)
-            .add(searchBar)
+            .searchable(text: $searchText)
+#if os(visionOS)
+            .onChange(of: sort) { _, value in
+                currentSymbols = value.symbols
+            }
+#else
             .onChange(of: sort) { value in
                 currentSymbols = value.symbols
             }
+#endif
             .toolbar{
                 ToolbarItem(placement: .primaryAction){
                     filterMenu
@@ -106,16 +116,11 @@ public struct SFSymbolPicker: View {
                 }
             }
         } label: {
-            if #available(iOS 15, *) {
-                Image(symbol: .filter)
-            } else {
-                Image(systemName: "line.horizontal.3.decrease.circle")
-            }
+            Image(symbol: .filter)
         }
     }
 }
 
-@available(iOS 14, *)
 struct SFSymbolPicker_Previews: PreviewProvider {
     static var previews: some View {
         SFSymbolPicker(symbolTitle: SFSymbol.share.title){ _ in}
