@@ -7,6 +7,6 @@
 
 import UIKit
 
-public extension UIColor{
+public extension UIColor {
     static let safari = UIColor(#colorLiteral(red: 0.1193895862, green: 0.706346333, blue: 0.9974301457, alpha: 1))
 }
