@@ -7,7 +7,7 @@
 
 import Foundation
 
-public extension URL{
+public extension URL {
     static let gitHub   = URL(string: "https://github.com/Rspoon3/RSWTools")!
     static let personal = URL(string: "https://www.rspoon3.com")!
 }
